+++ src/functions.php
@@ -1,12 +1,10 @@
 <?php
 namespace GuzzleHttp;
 
-use GuzzleHttp\Exception\InvalidArgumentException;
 use GuzzleHttp\Handler\CurlHandler;
 use GuzzleHttp\Handler\CurlMultiHandler;
 use GuzzleHttp\Handler\Proxy;
 use GuzzleHttp\Handler\StreamHandler;
-use Psr\Http\Message\UriInterface;
 
 /**
  * Expands a URI template
@@ -99,8 +97,8 @@ function debug_resource($value = null)
  *
  * The returned handler is not wrapped by any default middlewares.
  *
- * @throws \RuntimeException if no viable Handler is available.
  * @return callable Returns the best handler for the given system.
+ * @throws \RuntimeException if no viable Handler is available.
  */
 function choose_handler()
 {
@@ -333,62 +331,4 @@ function json_encode($value, $options = 0, $depth = 512)
     }
 
     return $json;
-}
-
-/**
- * Wrapper for the hrtime() or microtime() functions
- * (depending on the PHP version, one of the two is used)
- *
- * @return float|mixed UNIX timestamp
- * @internal
- */
-function _current_time()
-{
-    return function_exists('hrtime') ? hrtime(true) / 1e9 : microtime(true);
-}
-
-
-/**
- * @param int $options
- *
- * @return UriInterface
- *
- * @internal
- */
-function _idn_uri_convert(UriInterface $uri, $options = 0)
-{
-    if ($uri->getHost()) {
-        $idnaVariant = defined('INTL_IDNA_VARIANT_UTS46') ? INTL_IDNA_VARIANT_UTS46 : 0;
-        $asciiHost = $idnaVariant === 0
-            ? idn_to_ascii($uri->getHost(), $options)
-            : idn_to_ascii($uri->getHost(), $options, $idnaVariant, $info);
-        if ($asciiHost === false) {
-            $errorBitSet = isset($info['errors']) ? $info['errors'] : 0;
-
-            $errorConstants = array_filter(array_keys(get_defined_constants()), function ($name) {
-                return substr($name, 0, 11) === 'IDNA_ERROR_';
-            });
-
-            $errors = [];
-            foreach ($errorConstants as $errorConstant) {
-                if ($errorBitSet & constant($errorConstant)) {
-                    $errors[] = $errorConstant;
-                }
-            }
-
-            $errorMessage = 'IDN conversion failed';
-            if ($errors) {
-                $errorMessage .= ' (errors: ' . implode(', ', $errors) . ')';
-            }
-
-            throw new InvalidArgumentException($errorMessage);
-        } else {
-            if ($uri->getHost() !== $asciiHost) {
-                // Replace URI only if the ASCII version is different
-                $uri = $uri->withHost($asciiHost);
-            }
-        }
-    }
-
-    return $uri;
 }
